wip(#1102): extract emit/binary.rs + emit/method_call.rs
